intel-x86: Use prepend to set the "intel-x86" override
Issue: LINCD-5976, LIN1021-752

We have set the SRCREV_machine_intel-x86 to "AUTOREV" in BSP layer,
and set the SRCREV_machine_kb-dev-standard-intel-x86 to one specific
commit id in wrlinux layer. If both the "intel-x86" and
"kb-dev-standard-intel-x86" are set in MACHINEOVERRIDES, the last
one will take precedence. In this cases, we definitely want the one
set by the srcrev.inc take affect, so use the prepend to set the
"intel-x86" override to make sure that it is always before the
"kb-dev-standard-intel-x86" in the MACHINEOVERIDES.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>