intel-x86: config: remove setting WRL_RECIPES from templates
Issue: LINUXEXEC-23239

Some customers want access to use recipes that are included in
WRL_RECIPES but they do not use templates. By moving the updates to
WRL_RECIPES outside of the feature and default templates we can
simplify this usecase.

We move the changes to WRL_RECIPES to wrlinux-recipelists.conf or for
local recipes to the layer.conf. Since this is a Wind River layer we
can host the WRL_RECIPES here to save having to edit multiple layers
when we add or remove recipes to this layer.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>